fix: close the determinism class + v4 registered baselines
Codex's confirmation pass proved duplicate turn text (267/500 questions)
still let production's id-based search tie-order leak into simulation
top-K selection. Structural closure: a stable (score DESC, sessionIndex,
turnIdx) re-sort of the full result list at the single consumption
point, so no production tie-order can reach the harness. Proven on the
exact codex repro (gpt4_6dc9b45b: all 30 rounds identical across two
independent ingests, including the two rounds that diverged before).

Test suite split (harness + determinism files, shared fixtures module)
and generic pipeline tests trimmed to 8 simulation rounds so no vitest
worker crosses the Windows 60s RPC limit: exit 0 in ~18s, 17/17, two
consecutive runs.

v4 registered run: every reported cell identical to v3 at 4 decimal
places (heldout@0.3 recency 0.4203, strength 0.3049, uniform 0.2468) -
the fix added provable reproducibility without moving any aggregate.
