+++ benchmarks/memory-value/evidence-full500.log
@@ -1,12 +1,12 @@
 [run] REAL data=C:/Users/skf_s/hippo/benchmarks/longmemeval/data/longmemeval_s_cleaned.json questions=500/500 skipSimulate=false budget=0.3 seed=42
-[run] stages: ingest=2644835ms simulate=2136450ms extract=5844ms evaluate=1735ms total=4788869ms
+[run] stages: ingest=2692899ms simulate=2168840ms extract=5984ms evaluate=1854ms total=4869590ms
 [run] gold modes: {"evidence-turn":500,"answer-session-all":0}
 [run] variance gate: 8/30 features vary within at least one store (min required: 6)
 [run] varying: age_days, half_life_days, strength, retrieval_count, outcome_positive, outcome_negative, outcome_ratio, content_length
 [run] dead: schema_fit, pinned, starred, error_tag, tag_count, dag_level, valence_neutral, valence_positive, valence_negative, valence_critical, layer_buffer, layer_episodic, layer_semantic, layer_trace, kind_raw, kind_distilled, kind_superseded, kind_archived, confidence_verified, confidence_observed, confidence_inferred, confidence_stale
 [run] heldout: uniform@0.3=0.2468 recency@0.3=0.4203 (n=190, skipped0gold=10)
 [run] train: uniform@0.3=0.2426 recency@0.3=0.3862 (n=289, skipped0gold=11)
-[run] wrote C:\Users\skf_s\hippo-wt-lc2e1\benchmarks\memory-value\results\2026-08-09T10-52-43-269Z.json
+[run] wrote C:\Users\skf_s\hippo-wt-lc2e1\benchmarks\memory-value\results\2026-08-09T12-48-06-308Z.json
 [run] wrote C:\Users\skf_s\hippo-wt-lc2e1\benchmarks\memory-value\results\latest.json
-(node:110104) ExperimentalWarning: SQLite is an experimental feature and might change at any time
+(node:108252) ExperimentalWarning: SQLite is an experimental feature and might change at any time
 (Use `node --trace-warnings ...` to show where the warning was created)
