+++ benchmarks/memory-value/split-registered.json
@@ -543,6 +543,6 @@
     "fullTrainCount": 300,
     "fullHeldoutCount": 200
   },
-  "generatedAt": "2026-08-09T09:32:28.810Z",
+  "generatedAt": "2026-08-09T11:26:31.974Z",
   "source": "C:/Users/skf_s/hippo/benchmarks/longmemeval/data/longmemeval_s_cleaned.json"
 }
